Move Emoji to the utils package
Emoji is an utilitarian class.
