+++ src/main/java/ru/kingsbot/command/TextCommandParser.java
@@ -1,7 +1,7 @@
 package ru.kingsbot.command;
 
 import ru.kingsbot.Bot;
-import ru.kingsbot.Emoji;
+import ru.kingsbot.utils.Emoji;
 import ru.kingsbot.command.keyboard.Keyboards;
 import ru.kingsbot.entity.Player;
 import ru.kingsbot.entity.army.Clubman;
