+++ src/main/java/ru/kingsbot/command/group/BuildingCommand.java
@@ -1,6 +1,6 @@
 package ru.kingsbot.command.group;
 
-import ru.kingsbot.Emoji;
+import ru.kingsbot.utils.Emoji;
 import ru.kingsbot.command.Command;
 import ru.kingsbot.command.keyboard.Button;
 import ru.kingsbot.command.keyboard.Color;
