+++ src/main/java/ru/kingsbot/command/group/building/BuyBuildingCommand.java
@@ -1,6 +1,6 @@
 package ru.kingsbot.command.group.building;
 
-import ru.kingsbot.Emoji;
+import ru.kingsbot.utils.Emoji;
 import ru.kingsbot.command.Command;
 import ru.kingsbot.command.keyboard.Keyboards;
 import ru.kingsbot.entity.Player;
