+++ src/main/java/ru/kingsbot/command/group/market/ChoseResourceCommand.java
@@ -1,6 +1,6 @@
 package ru.kingsbot.command.group.market;
 
-import ru.kingsbot.Emoji;
+import ru.kingsbot.utils.Emoji;
 import ru.kingsbot.command.Command;
 import ru.kingsbot.command.keyboard.Button;
 import ru.kingsbot.command.keyboard.Color;
